[CHORE] Error 위치 변경 (#547)
